rename: 마이페이지 api 폴더 경로 수정
