--- src/apis/myPage/axios.ts
@@ -1,10 +1,0 @@
-import { privateInstance } from '@apis/instance';
-
-const getMyPage = async (userId: string) => {
-  const res = await privateInstance.get('/user/mypage', {
-    params: { userId },
-  });
-  return res.data;
-};
-
-export default getMyPage;
