--- src/apis/myPage/type.ts
@@ -1,9 +1,0 @@
-export interface MyPage {
-  userId: number;
-  nickname: string;
-  email: string;
-  ageRange?: string;
-  gender?: string;
-  religion?: string;
-  hasExperience?: boolean;
-}
