+++ src/apis/user/type.ts
@@ -9,3 +9,13 @@ export interface OnboardingUserRequest {
   religion: string | null;
   hasExperience: string | null;
 }
+
+export interface MyPageType {
+  userId: number;
+  nickname: string;
+  email: string;
+  ageRange?: string;
+  gender?: string;
+  religion?: string;
+  hasExperience?: boolean;
+}
